eclass-local-contextadmin
Updated the way category admin loads the context (moodle v2.2.3 changed the standard).  Setting the category now automatically sets the context and direct changed to the context should never be made.
Updated how the blocks.php in local/contextadmin loads the category in order to set the context for catadmin